The referenceTicketGroupId must be Int32, otherwise it returns "referenceTicketGroupId":{"reasons":["JSON integer 1884284108541005882 is too large or small for an Int32."]}.

+++ ticketnetwork-client-model/src/main/java/org/oxerr/ticketnetwork/client/model/TicketGroupV4PostModel.java
@@ -133,7 +133,7 @@ public class TicketGroupV4PostModel implements Serializable {
 	/**
 	 * The reference (external) ticket group ID.
 	 */
-	private Long referenceTicketGroupId;
+	private Integer referenceTicketGroupId;
 
 	/**
 	 * Flag indicating whether this is a short ticket group.
@@ -318,11 +318,11 @@ public class TicketGroupV4PostModel implements Serializable {
 		this.pending = pending;
 	}
 
-	public Long getReferenceTicketGroupId() {
+	public Integer getReferenceTicketGroupId() {
 		return referenceTicketGroupId;
 	}
 
-	public void setReferenceTicketGroupId(Long referenceTicketGroupId) {
+	public void setReferenceTicketGroupId(Integer referenceTicketGroupId) {
 		this.referenceTicketGroupId = referenceTicketGroupId;
 	}
